requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==4.0.0
appnope==0.1.3
asttokens==2.3.0
av==10.0.0
backcall==0.2.0
billiard==4.1.0
boto3==1.28.41
botocore==1.31.41
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
ctranslate2==3.19.0
decorator==4.4.2
distro==1.8.0
elevenlabs==1.2.2
executing==1.2.0
fastapi==0.103.0
faster-whisper==0.7.1
ffmpeg-python==0.2.0
filelock==3.12.3
flatbuffers==23.5.26
fsspec==2023.9.0
future==0.18.3
fuzzywuzzy==0.18.0
graphqlclient==0.2.4
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
ipyplot==1.1.1
ipython==8.15.0
jedi==0.19.0
jmespath==1.0.1
joblib==1.3.2
kombu==5.3.2
matplotlib-inline==0.1.6
Metaphone==0.6
moviepy @ git+https://github.com/Zulko/moviepy.git@bc8d1a831d2d1f61abfdf1779e8df95d523947a5
mpmath==1.3.0
nltk==3.8.1
numpy==1.25.2
onnxruntime==1.15.1
openai==1.33.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.2
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.0
regex==2023.10.3
replicate==0.13.0
requests==2.31.0
s3transfer==0.6.2
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
supervisor==4.2.5
sympy==1.12
tokenizers==0.13.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
websockets==12.0
